fix(core): additonal Fields with any getters
no inheritance for context objects. 
Using Any Getters and Setters for additional Fields.
